--- core/src/models/async/mod.rs
@@ -1,3 +1,0 @@
-mod unpack_task_data;
-
-pub use unpack_task_data::*;
