+++ core/src/models/mod.rs
@@ -3,6 +3,7 @@ mod file_entry;
 mod package_file_table;
 mod packed_package_buffer;
 mod path_tree;
+mod unpack_request;
 mod unpack_task_response;
 
 pub use benchmarked::*;
@@ -10,10 +11,5 @@ pub use file_entry::*;
 pub use package_file_table::*;
 pub use packed_package_buffer::*;
 pub use path_tree::*;
+pub use unpack_request::*;
 pub use unpack_task_response::*;
-
-#[cfg(feature = "async")]
-mod r#async;
-
-#[cfg(feature = "async")]
-pub use r#async::*;
